mac80211: add ATH9K_CHANNEL_CONTEXT to ath.mk
This is required for DBDC on ath9k. It needs to be enabled, right now,
it is disabled by default. It exposes ATH9K_CHANNEL_CONTEXT as option.

This could be interesting for Turris 1.x routers, which are using miniPCIe card - WNC DNXA-H1, there could be also others.

Also, it is necessary to add this peremeter to kernel
ath9k.ath9k_use_chanctx=1 to have this working.

Signed-off-by: Josef Schlehofer <pepe.schlehofer@gmail.com>