toolchain/gcc: introduce option GCC_POWERPC_SPE
PowerPC SPE support was removed in GCC 9 and newer releases because it was not maintained. [^1] [^2]

The usage of SPE[^3] ISA extension allows CPUs with PowerPC e500v2 cores (used in Turris 1.x routers, NXP P2020RDB
boards, ...) to perform native operations on single[^4] and double precision floats stored in GPRs[^5] as well
as vector operations using the whole 64-bit size of CPU registers (whose upper 32 bits are inaccessible otherwise).[^6]
This enables programs to be compiled for these CPUs without the need for floating-point arithmetic translation
leading to better performance.

[^1] https://www.phoronix.com/news/GCC-Removes-PowerPCSPE
[^2] https://gcc.gnu.org/git/?p=gcc.git;a=commitdiff;h=b31d0348ddada49453e3edaaf93a423fdc61dc79
[^3]: signal processing engine
[^4]: applies to e500v1 as well
[^5]: general purpose registers
[^6]: https://www.nxp.com/docs/en/reference-manual/SPEPEM.pdf

Signed-off-by: Josef Schlehofer <pepe.schlehofer@gmail.com>
[patch renumbered, commit message edited]
Signed-off-by: Šimon Bořek <simon.borek@nic.cz>